refactor validate number

+++ requests/account_requests.go
@@ -30,8 +30,8 @@ func (rr SearchAccountRequest) Validate() error {
 	return validation.ValidateStruct(&rr,
 		validation.Field(&rr.Limit, validation.Min(0)),
 		validation.Field(&rr.Offset, validation.Min(0)),
-		validation.Field(&rr.TimeFrom, validation.Min(0)),
-		validation.Field(&rr.TimeTo, validation.Min(0)),
+		validation.Field(&rr.TimeFrom, validation.Min(int64(0))),
+		validation.Field(&rr.TimeTo, validation.Min(int64(0))),
 		validation.Field(&rr.TimeTo, validation.By(checkTimeTimeFromTo(rr.TimeFrom, rr.TimeTo))),
 	)
 }
