+++ requests/order_request.go
@@ -23,8 +23,8 @@ func (rr SearchOrderRequest) Validate() error {
 		validation.Field(&rr.Limit, validation.Min(float32(0))),
 		validation.Field(&rr.Offset, validation.Min(float32(0))),
 		validation.Field(&rr.Type, validation.In(utils.IMPORT, utils.EXPORT)),
-		validation.Field(&rr.TimeFrom, validation.Min(0)),
-		validation.Field(&rr.TimeTo, validation.Min(0)),
+		validation.Field(&rr.TimeFrom, validation.Min(int64(0))),
+		validation.Field(&rr.TimeTo, validation.Min(int64(0))),
 		validation.Field(&rr.TimeTo, validation.By(checkTimeTimeFromTo(rr.TimeFrom, rr.TimeTo))),
 	)
 }
@@ -91,8 +91,8 @@ type ExportOrderRequest struct {
 func (rr ExportOrderRequest) Validate() error {
 	return validation.ValidateStruct(&rr,
 		validation.Field(&rr.Type, validation.In(utils.IMPORT, utils.EXPORT)),
-		validation.Field(&rr.TimeFrom, validation.Min(0)),
-		validation.Field(&rr.TimeTo, validation.Min(0)),
+		validation.Field(&rr.TimeFrom, validation.Min(int64(0))),
+		validation.Field(&rr.TimeTo, validation.Min(int64(0))),
 		validation.Field(&rr.TimeTo, validation.By(checkTimeTimeFromTo(rr.TimeFrom, rr.TimeTo))),
 	)
 }
